requirements.txt
click
gitpython
docker
requests